int -> size_t for number of devices return argument

+++ include/oc_core_res.h
@@ -409,9 +409,9 @@ int oc_core_get_latency(void);
 /**
  * @brief return the number of registered devices
  * 
- * @return int The number of registered devices
+ * @return size_t The number of registered devices
  */
-int oc_number_of_devices();
+size_t oc_number_of_devices();
 
 #ifdef __cplusplus
 }
